# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "One place to know about company events, activities and etc.",
        "version": "1.0.0",
        "title": "Intranet Portal",
        "contact": {
            "email": "sameer_shaik@merilytics.com"
        }
    },
    "host": "localhost:3008",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/employee/getEmployeeDetails": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "fetches entire org active employee details.",
                "description": "fetches entire org active employee details.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass pagination obj",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employeePagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/employeeDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/employee/getEmployeeDetailsById": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "fetches employee details by id.",
                "description": "fetches employee details by id.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass any employee id",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employeeId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/employeeDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/employee/getEmployeeDetailsWithToken": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "fetches employee details by id.",
                "description": "fetches employee details by id.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass any employee id",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/employeeDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/employee/getAppreciations": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "fetches employee details by id.",
                "description": "fetches employee details by id.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass any employee id",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appreciationPagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/employeeDetails"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/employee/getAppreciationsById": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "fetches appreciations details by id.",
                "description": "fetches appreciations details by id.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass any employee id",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employeeId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/appreciations"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/employee/sendAppreciations": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Appreciate employee",
                "description": "Appreciate employee",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appreciate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/employee/employeeSorting": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Employee Sorting drop down.",
                "description": "Employee Sorting drop down.",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SortingDropDown"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/employee/employeeSiloView": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Employee silo view",
                "description": "Employee silo view",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employeeId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/siloView"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
        "/employee/getEmployeeDetailsDropDown": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Employee silo view",
                "description": "Employee silo view",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "pass the specified fields in sample data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/employeeNameSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/siloView"
                            }
                        }
                    },
                    "404": {
                        "description": "Error occurred while fetching"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
        "employeeId": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "example": {
                "employeeId": 21258
            }
        },
        "employeePagination": {
            "type": "object",
            "properties": {
                "search": {
                    "type": "string"
                },
                "workLocationId": {
                    "type": "integer",
                    "format": "int64"
                },
                "filterId": {
                    "type": "integer",
                    "format": "int64"
                },
                "pageNo": {
                    "type": "integer",
                    "format": "int64"
                },
                "pageSize": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "example": {
                "search": null,
                "workLocationId": null,
                "sortById": null,
                "pageNo": 1,
                "pageSize": 10
            }
        },
        "appreciations": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "example": {
                "appreciationWallId": 5,
                "appreciationTo": 21258,
                "appreciationToName": "Sameer Shaik",
                "appreciatedBy": 21130,
                "appreciatedByName": "Sarath Kumardfgdfgdfg Gogineni",
                "contentOfAppreciation": "bla bla",
                "appreciatedTime": "2018-08-20T20:20:48.283Z",
                "createdBy": 21130,
                "createdByName": "Sarath Kumardfgdfgdfg Gogineni",
                "updatedBy": 21130,
                "updatedByName": "Sarath Kumardfgdfgdfg Gogineni",
                "createdDate": "2018-08-20T20:20:52.497Z",
                "updatedDate": "2018-08-20T20:20:38.267Z"
            }
        },
        "appreciate": {
            "type": "object",
            "properties": {
                "appreciatedBy": {
                    "type": "integer",
                    "format": "int64"
                },
                "appreciationTo": {
                    "type": "integer",
                    "format": "int64"
                },
                "contentOfAppreciation": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "integer",
                    "format": "int64"
                },
                "updatedBy": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "example": {
                "appreciatedBy": 21100,
                "appreciationTo": 21102,
                "contentOfAppreciation": "bla bla 222",
                "createdBy": 21130,
                "updatedBy": 21130
            }
        },
        "employeeDetails": {
            "type": "object",
            "properties": {
                "employeeId": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "example": {
                "employeeId": 21258,
                "employeeName": "Sameer Shaik",
                "dateOfJoining": "2017-02-20T00:00:00.000Z",
                "designation": "Analyst (IT)",
                "domain": "Technology",
                "technologiesKnown": "-",
                "reportingManager": "Sarath Kumar Gogineni",
                "workLocation": "KKR Square",
                "email": "sameer_shaik@merilytics.com",
                "skype": "sip:sameer_shaik@merilytics.com",
                "mail": "mailto:sameer_shaik@merilytics.com"
            }
        },
        "message": {
            "type": "object",
            "properties": {},
            "example": {
                "message": "Appreciations sent successfully."
            }
        },
        "SortingDropDown": {
            "type": "object",
            "properties": {},
            "example": [
                {
                    "sortById": 1,
                    "sortingName": "Employee ID"
                },
                {
                    "sortById": 2,
                    "sortingName": "Employee Name"
                },
                {
                    "sortById": 3,
                    "sortingName": "Date of Joining"
                }
            ]
        },
        "siloView": {
            "type": "object",
            "properties": {},
            "example": [
                {
                    "EmployeeID": 21100,
                    "employeeName": "Paavan Choudary",
                    "immediateReportingManagerID": null,
                    "designation": "CEO",
                    "emailID": "paavan@merilytics.com",
                    "dateOfJoining": "2011-01-01T00:00:00.000Z",
                    "EmpLevel": -2
                }
            ]
        },
...
        "appreciationPagination": {
            "type": "object",
            "properties": {},
            "example": {
                "employeeId": 21258,
                "pageNo": 1,
                "pageSize": 10
            }
        },
        "token": {
            "type": "object",
            "properties": {},
            "example": {
                "token": "353472ce8fad6c70"
            }
        },
        "employeeNameSearch": {
            "type": "object",
            "properties": {},
            "example": {
                "employeeId": 21258,
                "searchData": "s"
            }
        },
...
    },
...
}